wip: refactoring to python package
