+++ example/example/apps/polls/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class PollsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'backend.polls'
+    name = 'example.apps.polls'
