+++ example/example/controllers/test_login.py
@@ -1,7 +1,7 @@
 from fastapi.testclient import TestClient
 
 
-from frontend.app import app_controller
+from example.app import app_controller
 
 client = TestClient(app_controller.app)
