+++ example/example/main.py
@@ -1,0 +1,4 @@
+from example.app import controller
+
+# Expose for WSGI
+app = controller.app
