+++ example/example/views/package.json
@@ -9,12 +9,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@nivo/bar": "^0.87.0",
+    
     "postcss-cli": "^11.0.0",
+    "tailwindcss": "^3.4.1",
+    
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.52.0",
-    "tailwindcss": "^3.4.1"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.48",
