--- example/frontend/main.py
@@ -1,4 +1,0 @@
-from frontend.app import app_controller
-
-# Expose for WSGI
-app = app_controller.app
