--- example/frontend/views/app/home/page.tsx
@@ -1,118 +1,0 @@
-import * as React from "react";
-import {useServer} from "./_server";
-import {ResponsiveBar} from "@nivo/bar";
-
-const Home = () => {
-  const serverState = useServer();
-
-  return (
-    <div className="p-6">
-      <h1 className="text-2xl">Welcome &lt;{serverState.user?.username || "anon"}&gt;</h1>
-      <div>
-        <small>{serverState.questions.length} questions</small>
-        <ol>
-          {serverState.questions.map((question, idx) => (
-            <li key={question.pub_date}>{idx + 1}. {question.question_text}</li>
-          ))}
-        </ol>
-        <ResponsiveBar
-          data={[
-            {country: "AD", "hot dog": 13},
-            {country: "AE", "hot dog": 7},
-            {country: "AF", "hot dog": 9}
-          ]}
-          keys={["hot dog"]}
-          indexBy="country"
-          margin={{top: 50, right: 130, bottom: 50, left: 60}}
-          padding={0.3}
-          valueScale={{type: "linear"}}
-          indexScale={{type: "band", round: true}}
-          colors={{scheme: "nivo"}}
-          defs={[
-            {
-              id: "dots",
-              type: "patternDots",
-              background: "inherit",
-              color: "#38bcb2",
-              size: 4,
-              padding: 1,
-              stagger: true
-            },
-            {
-              id: "lines",
-              type: "patternLines",
-              background: "inherit",
-              color: "#eed312",
-              rotation: -45,
-              lineWidth: 6,
-              spacing: 10
-            }
-          ]}
-          fill={[
-            {
-              match: {
-                id: "fries"
-              },
-              id: "dots"
-            },
-            {
-              match: {
-                id: "sandwich"
-              },
-              id: "lines"
-            }
-          ]}
-          borderColor={{from: "color", modifiers: [["darker", 1.6]]}}
-          axisTop={null}
-          axisRight={null}
-          axisBottom={{
-            tickSize: 5,
-            tickPadding: 5,
-            tickRotation: 0,
-            legend: "country",
-            legendPosition: "middle",
-            legendOffset: 32
-          }}
-          axisLeft={{
-            tickSize: 5,
-            tickPadding: 5,
-            tickRotation: 0,
-            legend: "food",
-            legendPosition: "middle",
-            legendOffset: -40
-          }}
-          labelSkipWidth={12}
-          labelSkipHeight={12}
-          labelTextColor={{from: "color", modifiers: [["darker", 1.6]]}}
-          legends={[
-            {
-              dataFrom: "keys",
-              anchor: "bottom-right",
-              direction: "column",
-              justify: false,
-              translateX: 120,
-              translateY: 0,
-              itemsSpacing: 2,
-              itemWidth: 100,
-              itemHeight: 20,
-              itemDirection: "left-to-right",
-              itemOpacity: 0.85,
-              symbolSize: 20,
-              effects: [
-                {
-                  on: "hover",
-                  style: {
-                    itemOpacity: 1
-                  }
-                }
-              ]
-            }
-          ]}
-          animate={true}
-        />
-      </div>
-    </div>
-  );
-};
-
-export default Home;
